ingestion/requirements.txt
requests==2.31.0
azure-storage-blob==12.19.0
azure-identity==1.15.0
python-dotenv==1.0.0
pyodbc==5.0.1
